requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==5.0.7
idna==2.10
joblib==1.0.1
jstyleson
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
matplotlib==3.4.1
numba==0.53.1
numpy==1.20.2
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pooch==1.3.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
resampy==0.2.2
schema==0.7.4
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Shapely==1.7.1
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
urllib3==1.26.4
